More fiddling around. Fix a bug introduced when migrated to ARC.

+++ Classes/AILinkedList.h
@@ -14,7 +14,7 @@
 	AILinkedListObject *front, *back;
 }
 /* Returns the object */
-- (id)objectAtIndex:(unsigned)index; /* O(N) */
+- (id)objectAtIndex:(NSInteger)index; /* O(N) */
 - (id)objectAtFront; /* O(1) */
 - (id)objectAtEnd; /* O(1) */
